Переход от double к float. Использование FLT_EPSILOP

+++ Lab_4/ShapeTest/VerifyPoint.h
@@ -2,4 +2,4 @@
 #include "stdafx.h"
 #include "..\Shape\Point.h"
 
-void VerifyPoint(CPoint const& point, double expectedX, double expectedY);
+void VerifyPoint(CPoint const& point, float expectedX, float expectedY);
